Start documenting yum/dnf repo updateinfo.xml metadata
For `updateinfo.xml` specifically, it is currently parsed by many third
party tooling for the purpose of gathering information about updates
applicable to Linux distributions using `yum`/`dnf` repositories. Since
the format of `updateinfo.xml` has not been well defined over the years,
this presents challenges to authors of these tools.

With clear documentation on what exists in the wild, all authors of
tools producing or parsing `updateinfo.xml` can refer to one canonical
place for what exists in the wild and how to process it.

A TODO item is to develop a transitional and strict schema to help
producers of updateinfo.xml ensure compatibility with existing
tools (yum/dnf).

Signed-off-by: Stewart Smith <trawets@amazon.com>